feat(i18n): share editor i18n core with the portal
Extracts the TOML backend and language metadata/helpers into @shared/i18n so
the editor and portal use one system (same backend, language list, TOML
locale layout, Crowdin pipeline). The editor now imports + re-exports the
shared pieces (consumers unchanged); the portal builds its own instance from
the shared core with en-US as the source of truth. Converts the Users view as
the first migrated surface.
